Added caching mechanism

+++ inc/warehouse_db.h
@@ -10,7 +10,7 @@ void newConnectionCheck(void);
 void checkRequest(void);
 void processRequest(int client, char* msg);
 void cleanUp(void);
-void list(int id);
+void list(long int id);
 void listAll(void);
 void print_entry(int i);
 void dump(void);
